% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_d.R
\name{calculate_d}
\alias{calculate_d}
\title{Calculates a vector of KS values for each compound}
\usage{
calculate_d(cmpd, neg)
}
\arguments{
\item{cmpd}{list}

\item{neg}{dataframe or matrix}
}
\value{
D_values Dataframe
}
\description{
Given a dataframe for the negative control and a list of dataframes for each 
compound and concentration, this will calculate a signed D-statistic from a 
KS test for each feature (per compound and concentration).
}

