% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_metadata.R
\name{construct_metadata}
\alias{construct_metadata}
\title{Constructs metadata list to be used in functions}
\usage{
construct_metadata(df, compound_col, conc_col, feature_cols, negative_control)
}
\arguments{
\item{df}{Dataframe}

\item{compound_col}{index or name of dataframe column}

\item{conc_col}{index or name of dataframe column}

\item{feature_cols}{Vector of indices that match numeric feature columns}

\item{negative_control}{string, name of negative control in compound_col}
}
\value{
metadata list containing all metadata
}
\description{
Creates a list containing all necessary metadata to subset a dataframe
}

