% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cor_trim.R
\name{cor_trim}
\alias{cor_trim}
\title{Correlation between two vectors shifted by n}
\usage{
cor_trim(a, b, n = 0, shift = "left")
}
\arguments{
\item{a}{vector}

\item{b}{vector}

\item{n}{number of elements to shift the vector by}

\item{string,}{the direction to shift the first vector in.}
}
\value{
cor_out the correlation
}
\description{
Measures the correlation between two vectors shifted by 'n' elements. For
example, \code{cor_trim(a, b, n = 5, shift = 5)} will return the correlation
between the vectors a and b, with the last 5 elements of a removed, and the
first 5 elements of b removed
}
\examples{
a <- rnorm(100)
b <- rnorm(100)
cor_trim(a, b, n = 5)
cor_trim(a, b, n = 5, shift = "right")
}

