% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{euclid_dist}
\alias{euclid_dist}
\title{euclidean distance between two vectors}
\usage{
euclid_dist(a, b, ...)
}
\arguments{
\item{a}{numeric vector.}

\item{b}{numeric vector}

\item{...}{additional arguments to pass to \code{dist}}
}
\value{
Euclidean distance between the two vectors
}
\description{
as \code{dist()} has stupid defaults
}

