% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{example data}
\format{dataframe containing column per feature, row per observation}
\source{
Scott Warchal
}
\usage{
ex_data
}
\description{
example data from a typical CellProfiler screen
}
\keyword{datasets}

