% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_data.R
\name{get_compound_data}
\alias{get_compound_data}
\title{Subset compound data}
\usage{
get_compound_data(df, metadata)
}
\arguments{
\item{df}{dataframe}

\item{metadata}{list}
}
\value{
split_featuredata list
}
\description{
Creates a list of list of dataframes containing only feature data.
First splits df into a list, with each element containing a compound.
Then, splits each of the compound lists into a list of dataframes, with a 
dataframe corresponding to each compound-concentration pair.
}

