% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_negative_control.R
\name{get_negative_control}
\alias{get_negative_control}
\title{Creates dataframe of negative control featuredata}
\usage{
get_negative_control(df, metadata)
}
\arguments{
\item{df}{dataframe}

\item{metadata}{list}
}
\value{
df_subset dataframe
}
\description{
Given a dataframe and a metadata list, this function will subset the
dataframe and return only the featuredata for compounds matching the name
of the negative control
}

