% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_cols.R
\name{ks_cols}
\alias{ks_cols}
\title{Signed KS test across pairs of columns}
\usage{
ks_cols(f, g)
}
\arguments{
\item{f}{dataframe or matrix of sample}

\item{g}{dataframe or matrix of control}
}
\value{
tmp vector of D-statistics
}
\description{
Given two numerical columns of matrices, this function will compute a signed
D-statistic for every pair of columns.
}

