% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks_test}
\alias{ks_test}
\title{Signed Kolmogorov-Smirnov Test}
\usage{
ks_test(f, g)
}
\arguments{
\item{f}{Numerical vector to be compared to g}

\item{g}{Numerical vector, reference or control}
}
\value{
D D-statistic
}
\description{
A two-sample Kolmogorov-Smirnov test that also returns the direction of the
calculated supremum relative to the distribution g. Used within \code{ks_cols}
For example, if the values of g are smaller than g, and the CDF of f is
shift to the left of g, then the D statistic will be negative, if it is
shift to the right of g, then it will be positive.
}
\examples{
data(iris)
ks_test(iris[,1], iris[,2])

}

