% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cor.r
\name{loo_cor}
\alias{loo_cor}
\title{Leave one out correlation}
\usage{
loo_cor(x, i = 1)
}
\arguments{
\item{x}{matrix or dataframe of numerical vectors}

\item{i}{margin over which to divide the vectors. For example, if 1 is given
then this will separate the vectors per each row, if 2 is given, then the 
vectors will be separated column-wise}
}
\value{
out vector of correlation values, each value
}
\description{
A way to calculate the correlation between multiple vectors. This is a
quicker method that compares the correlation between a single vector any
the average for all the others.
}
\examples{
# correlation between rows of iris
loo_cor(iris[, 1:4])
# correlation between columns of iris
loo_cor(iris[, 1:4], 2)
}

