% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_shift.R
\name{max_shift}
\alias{max_shift}
\title{Find the shift with the best correlation}
\usage{
max_shift(cor_measurements, value = FALSE)
}
\arguments{
\item{cor_measurements}{vector of correlation values and shift}

\item{value}{boolean, whether to return a numerical correlation value}
}
\value{
max_shift shift with the highest correlation. If \code{value} is
	\code{TRUE} then it will also return the correlation value
}
\description{
For a number of correlation meansurements produced by \code{trim_cor}
will return the shift with the highest correlation value.
}
\examples{
# create dummy dataset
l <- list()
for (i in 1:100) l[[i]] <- rnorm(100)
df <- as.data.frame(l)
names(df) <- paste("col", 1:100, sep = "_")
# order values (more obvious correlation shifts)
df <- apply(df, 2, sort)
out <- trim_cor(df[,1], df[,2], n = 5, p = 5)

max_shift(out)
max_shift(out, value = TRUE)
}

