% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/max_trim.R
\name{max_trim}
\alias{max_trim}
\title{Converts output from max_shift() to a shifted vector
Trims a vector from the results of max_shift}
\usage{
max_trim(vector, max_shift)
}
\arguments{
\item{vector}{Vector to be trimmed}

\item{max_shift}{Character containing direction and number of elements to
        trim in that direction, or '0' for no trim. e.g \code{'+2'}.}
}
\value{
vector_out The trimmed vector
}
\description{
Given a vector, this function will trim it according to the output from
\code{max_trim}. That is will be trimmed to maximise correlation with the
rest of the vectors.
}
\examples{
# example data
v <- seq(1, 10, 1)

max_trim(v, '+2')
max_trim(v, '-3')
max_trim(v, '0')
}

