% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_trim.R
\name{max_trim_single}
\alias{max_trim_single}
\title{Converts output from max_shift() to a shifted vector
Trims a vector from the results of max_shift}
\usage{
max_trim_single(vector, max_shift)
}
\arguments{
\item{vector}{Vector to be trimmed}

\item{max_shift}{Character containing direction and number of elements to
trim in that direction, or '0' for no trim. e.g \code{'+2'}.}
}
\value{
vector_out The trimmed vector
}
\description{
Given a vector, this function will trim it according to the output from
\code{max_trim_single}. \code{max_trim_single} will trim directly the number
of elements specified by \code{max_shift}, with no input from the number of
features. In order to trim the vector by the number of features present
\code{max_trim} should be used instead of \code{max_trim_single}.
}
\examples{
# example data
v <- seq(1, 10, 1)

max_trim_single(v, '+2')
max_trim_single(v, '-3')
max_trim_single(v, '0')
}

