% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trim_cor.R
\name{plot_trim_cor}
\alias{plot_trim_cor}
\title{plot results from trim_cor}
\usage{
plot_trim_cor(x, ...)
}
\arguments{
\item{x}{a vector of correaltion values from \code{trim_cor}}

\item{...}{arguments to be passed to plot}
}
\value{
graph to the device
}
\description{
Given a vector of correlation values from \code{trim_cor}, this will produce
a line graph showing correlation with +/- shifts
}
\examples{

# create dummy dataset
l <- list()
for (i in 1:100) l[[i]] <- rnorm(100)
df <- as.data.frame(l)
names(df) <- paste("col", 1:100, sep = "_")
# order values (more obvious correlation shifts)
df <- apply(df, 2, sort)
out <- trim_cor(df[,1], df[,2], n = 10, p = 5)
plot_trim_cor(out, main = "plot of correlation values")
}

