% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity_list}
\alias{similarity_list}
\title{similarity score}
\usage{
similarity_list(x, return_matrix = TRUE, ...)
}
\arguments{
\item{x}{list of compound vectors produced by \code{trim()}}

\item{return_matrix}{if TRUE will return a distance matrix, otherwise a
dataframe will be returns}

\item{...}{additional arguments to pass to distance function, see \code{dist}}
}
\value{
distance matrix or dataframe, dependent on \code{return_matrix}
}
\description{
Essentially a pair-wise euclidean distance between each vector in the list
}

