% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiss.R
\name{tiss}
\alias{tiss}
\title{Titration invariant similarity score}
\usage{
tiss(data, metadata, align = TRUE)
}
\arguments{
\item{data}{input data in the form of a data.frame}

\item{metadata}{metadata object generated from /code{construct_metadata}}

\item{align}{whether to align compounds by maximum correlation in order
to account fo differences in potency. If FALSE will not align the
compounds and generate similarity scores between complete compound
vectors}
}
\value{
similarity matrix or datarame if \code{return_matrix = FALSE}
   is passed as an additional argument
}
\description{
Calculates the TISS between compounds. The function is a wrapper
for a typical workflow. Though it's generally recommended to run
the functions separately for ease of debugging and sanity checks
}

