% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim compound vectors}
\usage{
trim(x, max, metadata)
}
\arguments{
\item{x}{a scaled compound vector produced from \code{scale_d()}}

\item{max}{a list of maximum shifts, produced by the output from
\code{correlate} and \code{max_shift}}
}
\value{
a list of trimmed compound vectors
}
\description{
Given a scaled compound vector and list of shifts that maximise correlation
\code{trim} will trim each compound vector according to those shifts. i.e
they will be trimmed by n titrations worth of features, where n is the shift
}

