% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{trim_cor_}
\alias{trim_cor_}
\title{Internal function used in correlate()}
\usage{
trim_cor_(x, y, n, p)
}
\arguments{
\item{x}{First cor_vector}

\item{y}{Second vector}

\item{n}{Number of shifts in either direction}

\item{p}{Number of features per shift}
}
\value{
cor_measurements
}
\description{
\code{trim_cor} should be used interactively.
}

