% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Z-score of KS values}
\usage{
zscore(x)
}
\arguments{
\item{x}{Vector of D values}
}
\value{
out Vector of z-scored values
}
\description{
Used within \code{calculate_d}
Scale and center the D statistic produced by \code{ks_test}
}

