% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_ks.R
\name{zscore_ks}
\alias{zscore_ks}
\title{Z-score of KS values}
\usage{
zscore_ks(x)
}
\arguments{
\item{x}{Vector of D values}
}
\value{
out Vector of z-scored values
}
\description{
Scale and center the D statistic produced by \code{ks_test}
}

