% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/construct_metadata.R
\name{construct_metadata}
\alias{construct_metadata}
\title{Constructs metadata list to be used in functions}
\usage{
construct_metadata(df, compound_col, conc_col, feature_cols, negative_control)
}
\arguments{
\item{df}{Dataframe}

\item{compound_col}{Vector or dataframe column}

\item{conc_col}{Vector or dataframe column}

\item{feature_cols}{Vector}

\item{negative_control}{string}
}
\value{
metadata list containing all metadata
}
\description{
Creates a list containing all necessary metadata to subset a dataframe
}

