% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Calculates correlation between vectors shifted by titrations.}
\usage{
correlate(x, metadata)
}
\arguments{
\item{x}{Matrix or dataframe of compound vectors. Each column representing
an individual compound vector. This should have been produced by}

\item{metadata}{corresponding metadata for x, produced by \code{calculate_d}
and scaled with \code{scale_d}
\code{construct_metadata}}
}
\value{
List of correlation measurements, each element of the list
  corresponds to a compound vector in the order given by \code{x}
}
\description{
For a matrix, with each row representing a compound vector constructed of
features and titrations. \code{correlate} will calculate the correlation 
for each vector compared against the average of all the other vectors,
shifted by a number of titrations in both directions. This is to align
compound vectors invariant of differing potencies.
}

